Move auth lib into its own folder
